+++ lib/mongoid/document.rb
@@ -14,64 +14,6 @@ module Mongoid
       Mongoid.register_model(self)
     end
 
-    # Default comparison is via the string version of the id.
-    #
-    # @example Compare two documents.
-    #   person <=> other_person
-    #
-    # @param [ Document ] other The document to compare with.
-    #
-    # @return [ Integer ] -1, 0, 1.
-    #
-    # @since 1.0.0
-    def <=>(other)
-      attributes["_id"].to_s <=> other.attributes["_id"].to_s
-    end
-
-    # Performs equality checking on the document ids. For more robust
-    # equality checking please override this method.
-    #
-    # @example Compare for equality.
-    #   document == other
-    #
-    # @param [ Document, Object ] other The other object to compare with.
-    #
-    # @return [ true, false ] True if the ids are equal, false if not.
-    #
-    # @since 1.0.0
-    def ==(other)
-      self.class == other.class &&
-        attributes["_id"] == other.attributes["_id"]
-    end
-
-    # Performs class equality checking.
-    #
-    # @example Compare the classes.
-    #   document === other
-    #
-    # @param [ Document, Object ] other The other object to compare with.
-    #
-    # @return [ true, false ] True if the classes are equal, false if not.
-    #
-    # @since 1.0.0
-    def ===(other)
-      other.class == Class ? self.class === other : self == other
-    end
-
-    # Delegates to ==. Used when needing checks in hashes.
-    #
-    # @example Perform equality checking.
-    #   document.eql?(other)
-    #
-    # @param [ Document, Object ] other The object to check against.
-    #
-    # @return [ true, false ] True if equal, false if not.
-    #
-    # @since 1.0.0
-    def eql?(other)
-      self == (other)
-    end
-
     # Freezes the internal attributes of the document.
     #
     # @example Freeze the document
@@ -147,6 +89,7 @@ module Mongoid
       _building do
         @new_record = true
         @attributes ||= {}
+        @attributes_before_type_cast ||= {}
         options ||= {}
         apply_pre_processed_defaults
         process_attributes(attrs, options[:as] || :default, !options[:without_protection]) do
@@ -342,6 +285,7 @@ module Mongoid
         doc = allocate
         doc.criteria_instance_id = criteria_instance_id
         doc.instance_variable_set(:@attributes, attributes)
+        doc.instance_variable_set(:@attributes_before_type_cast, {})
         doc.apply_defaults
         IdentityMap.set(doc) unless _loading_revision?
         yield(doc) if block_given?